Fix React bugs, prep for key diagraming
React bugs:
Set React key property in the list of Key components inside Keyboard,
not (improperly) on the non-list <button> element inside Key.
Base the React key for Key component on location of key in Keyboard.

Key diagraming:
I want to draw lines to keys on the keyboard inside of KeyInfo.
Set id for Key components inside the Keyboard to support this
in the future.