Take it back: we do guarantee a default keyboard
Add a default keyboard for showing an error

This is much easier to work with than having a special case for no
board defined. Instead of having a "title" board like before though,
we'll have a "default" board, or maybe later we'll rename it to an
"error" board, which will show the user that there's a configuration
problem.